Changed game ids from int to int64 just in case...

+++ manifest/manifest-file-iterator.go
@@ -3,7 +3,7 @@ package manifest
 import "errors"
 
 type FileInfo struct {
-	GameId   int
+	GameId   int64
 	Kind     string
 	Name     string
 	Checksum string
